+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.0.3] - 2019-01-31
 ### Added
 - Chat, instrument, apikey resources
 ### Changed
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Stats, settlement, schema, quote, position resources
 - Orderbook, order resources
 - Liquidation, leaderboard, insurance, execution, user, announcement
+- Make Websocket API interface generic
 
 ## [0.0.2] - 2019-01-22
 ### Added